refactor(CtRolePathElement): Replace old behavior by new one from CtUniqueRolePathElement
